+++ src/instrumentation.ts
@@ -1,7 +1,9 @@
 export async function register() {
   if (process.env.NEXT_RUNTIME === 'nodejs') {
-    const { checkLicenses, checkSubscriptions } = await import('./cronTasks')
+    const { checkLicenses, checkSubscriptions, processRecurringInvoices, cleanupPortalSessions } = await import('./cronTasks')
     checkLicenses()
     checkSubscriptions()
+    processRecurringInvoices()
+    cleanupPortalSessions()
   }
 }
